requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.3.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
python-dateutil>=2.8.0
pytz>=2023.3
